third_party/WebKit/LayoutTests/http/tests/push_messaging/pushmessagedata-expected.txt

This is a testharness.js-based test.
PASS PushMessageData initialization and accessor behaviour in a ServiceWorker.
PASS PushEvent can be initialized from ArrayBuffer, ArrayBufferView and USVStrings.
PASS PushMessageData is null by default.
PASS PushMessageData handling of ArrayBuffer content.
PASS PushMessageData handling of Blob content.
PASS PushMessageData handling of valid JSON content.
FAIL PushMessageData handling of invalid JSON content.Test bug: need to pass exception to assert_throws()
FAIL PushMessageData should not be constructable.Test bug: need to pass exception to assert_throws()
PASS PushEventInit data is not normalized
FAIL PushMessageData throws when passed SharedArrayBuffer view.Test bug: need to pass exception to assert_throws()
Harness: the test ran to completion.
